Revert "devfreq: Fix type of frequency table"
This reverts commit 5182cf7d50af37c54e84eb6c0b5a0a91588f564e.

+++ include/linux/devfreq.h
@@ -52,7 +52,7 @@ struct devfreq_dev_profile {
 	int (*get_cur_freq)(struct device *dev, unsigned long *freq);
 	void (*exit)(struct device *dev);
 
-	unsigned long *freq_table;
+	unsigned int *freq_table;
 	unsigned int max_state;
 	const struct devfreq_governor_data *governor_data;
 	unsigned int num_governor_data;
